Score Counter and candy count
Score count at the top of the screen as well as a global candy counter

+++ BooGame/Assets/Scripts/GameController.cs
@@ -7,7 +7,7 @@ public class GameController : MonoBehaviour
 
     public static GameController control;
 
-    public int candyCount;
+    public double candyCount;
 
     public float timeSinceDash;
     public float timeSinceLine;
